Added scripts to add users to multiple courses
Currently there are separate scripts for adding a user as a student
for multiple course and for adding a user as a grutor for multiple
courses. They are the same except for the second last line where the
field "courseStudent" or "courseGrutor" is modified. The code is based
on instructions from the user manual but has not yet been run
successfully.