Placeholder submission no longer created when grading all submissions
Affects giveFullPoints.py, giveSpecificScore.py and
markSubmissionsAsDone.py
Before, if there was not a submission, one would be created to be graded
with all the others. However, this may not be expected -- it seems odd
to give full credit if there was no submission. Now, if there is no
submission that user will just be skipped.