Added getSubmission.py
Note: this version has not been tested because the files are not
on my local machine and cannot be accessed. Not sure if it works
with the knuth file system either.
This script should ideally allow the user to get the actual student
submission files by specifying the assignment and problem.