Updated hw1pr2 tests: more informative test names and changed inputs
Since the submission site interface for setting up the tests just grabs the
test name, the names are more informative so it is clearer what each test is
meant for
Before the inputs that were used were the examples given to the students in
the directions, which will probably have been run as they develop. The values
have been changed now to hopefully be more useful as a test.