Reorganized test files: grouped by course not by test type
This will be helpful when there are test files for more than one course
(so as not to get confused which "hw0" goes with which course for example).
Also it is helpful to have all the test files for one course in one place so
they are easier to find. (Previously, an instructor might miss the CS5 hw12
files for example, because they were in a different directory than the
majority of the other test files.) Unfortunately the type of each test is a
bit less clear now, but the extensions should hopefully inform the user.