Added a script for creating a backup of Cicero's db remotely
To use this script, first put the admin password into the file.
Then, make the file executable and run it from the repo directory:
$ chmod +x mongo_backup.sh
$ bash mongo_backup.sh
Current version has locks to prevent the db from changing while the backup is
made -- this might not be ideal in the long run.
Next step is to figure out how to make this work with cron and then
to automatically upload the backups to Amazon (included in comments here but
not yet attempted)