Courses no longer added multiple times for courseStudent or courseGrutor
Before, in addUsersFromList.py, the course would always just be appeneded to
the existing list for the user (courseStudent or courseGrutor as was selected
earlier in the script). Now, there is a check to see if the course is already
in the list, and if so it is not added again.