Replaced reset password button with a message to email the instructor
Since the reset password function does not actually work yet, this
button causes confusion. The code for the button is left in comments
in the case that this function is fixed in the future. The acutal
function is in app/userViews/common/accounts.py and is called
requestRecovery() -- that function has not been modified at all.