Moved styling to separate css files, fixed bug with navbar resizing
Before, there were a few instances where styling was added using <style>
tags in the <head> element. These are hard to find and makes maintainance
and extension hard, so this code has been moved to separate files.
In addition, there was formerly a problem when the viewing window was
resized (specifically in widths 768-991px) where the navbar would become
too tall and cover some content on the page. This was due to not adjusting
the padding for the body correctly for those widths. This has now been
corrected with the styles inside of base-style.css