Racket autograder: totalTests fixed, changed expected test format
Before, the number of tests run would always appear to be 0. The way the tests
were structured, there was only information output by failed tests, and so
there was no way to get the total number of tests. Now, by using a test-suite
and the command "run-tests" in normal mode, we are able to get this
information. It is important to note that old test files will no longer run
properly though!