Racket autograder: totalTests works with old file format, more errors printed
Last commit, the expected file format for racket test files was changed
slightly (using a test-suite and the "run-tests" command) so that summary
information was available. However, the old files did not have that
information in the output and so would cause errors and not run at all.
Now the totalTests is just the empty string if that information is not
available (if the test file is in the old format).

Additionally, more information is printed to the celery worker window if
there is a timeout in the test file.