Racket autograder: fixed how returncode was assigned and used
In command.py before, there was a variable called status that was declared but
never used, and a variable called returncode that was not declared but was
used. Referencing sources for this file, it seems that "status" was renamed
to "returncode" but not in all places, so that change is now made.
In rackunit.py, the returncode is referenced to know if the test died or not.
The line using it is now more specific to match the error code set inside
command.py in this case.