Student View Grades Page reformatted: Vertical table layout
Before, there was a horizontal table layout that was difficult to
view and so therefore not very helpful. It has been restructured
based off of the instructor viewStudent.html template. There is
no longer the same highlighting because that came from the use
of the studentRenderGrades() function inside viewGrades.py (which
formats grades as all in a row), but this seems like a small trade off