Appearance: Added a button to get back to all assignments from single problem
When grading, graders select the course from the drop down "Grader" menu,
which takes them to a page where all the assignments for the course are listed
(file is app/templates/grutor/assignments.html). Then, they can select to grade
an individual problem which opens a different page with all the student submissions
listed (file is app/templates/grutor/problems.html). Before, there was not a good
way to get from the problem page to the assignment page. Now, a button has been
added so graders don't have to use the drop down menu again.