Appearance: First trial of student viewEuros page
This first pass has the euros displayed in a horizontal table (green boxes
indicate a submission and no euro used, red checks indicate a euro was used
because that submission was late). Euros are counted as being for a whole
assignment (so multiple late parts in one assignment are okay). This view
is an idea of how an instructor viewEuros page might look -- not meant to
be final for the student view. Improvements include: making the table
vertical, and possibly not breaking the detail down into individual problems