Student, Grutor and Instructor tables now sortable by header on course settings page
When these tables are created originally, the users are not displayed in any
logical order. Now, the tables can be sorted after they are created by clicking
on the "Username" header. Known bugs: the other headers "View Details" and
"Remove" are also clickable, and the resulting order has no discernable pattern