Removed "Make Page" buttons from problem settings page
Before, the instructor would have the option to make a wiki page for
the problem description and grading notes. Now, the option to link
pages still exists, but the links should be for external sites. Note
that if an internal wiki link is used, it will still work, but it is
harder now to create a url for that page.