Removing wiki as default for course homepage
Creating a new course at admin dashboard used to automatically include
a call that created an internal wiki page. Now, that line is commented
out, so the default on creation is just the character '#' which will
redirect to the same page. This default behavior is caused by the
properties set inside of app/structures/models/course.py, specifically
"homepage = db.StringField(default="#")" (line 293 at the time of this
commit)