Don't hardcode path to racket executable
The subprocess module can resolve executables on the $PATH. Therefore,
there is no need to provide the full path (/usr/bin/racket) to the
racket executable. In fact, doing so will break the autograder on
systems where racket is not installed to /usr/bin/racket (for instance,
on my system it is installed to /usr/local/bin/racket).