Normalize line endings and add .gitattributes
This is a cross-platform project. The standard line endings used in
cross-platform projects are LF, not CRLF.

Adding .gitattributes ensures that while line endings are LF in the
repository, they will silently be converted to CRLF when checked out on
Windows machines, and back to LF when checked in on Windows machines.