Improve error reporting in command.py
Previously, if an error occurred when constructing the Popen object,
Command.run would throw its own exception because self.process would not
be defined. This commit makes a small modification so that the traceback
from the error is instead returned from Command.run, which will no
longer throw an exception in this situation.